+++ src/main/src/main/java/org/geoserver/catalog/CatalogFacade.java
@@ -791,6 +791,6 @@ public interface CatalogFacade {
      *         filter
      */
     public <T extends CatalogInfo> CloseableIterator<T> list(final Class<T> of,
-            final Filter filter, @Nullable Integer offset, @Nullable Integer count,
+            final Filter filter, @Nullable Long offset, @Nullable Long count,
             @Nullable SortBy... sortOrder);
 }
